Fix vim installing script

+++ scripts/install_nvim.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
+if [ -d "/opt/nvim-linux64" ]; then
+    echo "Neovim is already installed!"
+    exit 0
+fi
+
 curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-sudo rm -rf /opt/nvim
+sudo rm -rf /opt/nvim-linux64
 sudo tar -C /opt -xzf nvim-linux64.tar.gz
 rm nvim-linux64.tar.gz
